+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/EdgeLord.cs
@@ -23,6 +23,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
             var FriendlyAuraFeature = Resources.GetModBlueprint<BlueprintFeature>("FriendlyAuraFeature");
             var TrueMainCharacter = Resources.GetModBlueprint<BlueprintFeature>("TrueMainCharacter");
             var OverpoweredAbilitySelection2 = Resources.GetModBlueprint<BlueprintFeatureSelection>("OverpoweredAbilitySelection2");
+            var CharacterDevelopmentSelection = Resources.GetModBlueprint<BlueprintFeatureSelection>("CharacterDevelopmentSelection");
 
             // Archetype
             var EdgeLordArchetype = Helpers.CreateBlueprint<BlueprintArchetype>("EdgeLordArchetype", bp => {
@@ -38,7 +39,9 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
                     Helpers.LevelEntry(8, IsekaiFastMovement),
                     Helpers.LevelEntry(9, FriendlyAuraFeature),
                     Helpers.LevelEntry(10, OverpoweredAbilitySelection2),
+                    Helpers.LevelEntry(13, CharacterDevelopmentSelection),
                     Helpers.LevelEntry(15, OverpoweredAbilitySelection2),
+                    Helpers.LevelEntry(17, CharacterDevelopmentSelection),
                     Helpers.LevelEntry(20, TrueMainCharacter),
                 };
                 bp.AddFeatures = new LevelEntry[] {
@@ -47,7 +50,8 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
                     Helpers.LevelEntry(7, EdgeLordFastMovement),
                     Helpers.LevelEntry(8, CripplingStrike),
                     Helpers.LevelEntry(10, ExtraStrike, DispellingAttack),
-                    Helpers.LevelEntry(15, ExtraStrike),
+                    Helpers.LevelEntry(11, CharacterDevelopmentSelection),
+                    Helpers.LevelEntry(15, ExtraStrike, CharacterDevelopmentSelection),
                     Helpers.LevelEntry(20, ExtraStrike),
                 };
                 bp.OverrideAttributeRecommendations = true;