+++ IsekaiMod/Content/Classes/IsekaiProtagonist/IsekaiProtagonistClass.cs
@@ -30,7 +30,8 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist
 
         public static void Add()
         {
-            // TODO: re-arrange character development feats (esp. villain and edge lord)
+            // TODO: add AddToSelection method to op ability selection
+            // TODO: add Hero archetype
 
             // TODO: add scaling natural armor, strength and dexterity (like animal companion) to deathsnatcher
             // TODO: add vampiric drain spell for isekai vampire heritage