+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added `Corrupt Aura` feature for `Villain` archetype.
+- Added `Exceptional feats` within bonus feat selection.
+- Added `Effect Immunity` exceptional feat.
 ### Changed
+- Reworked `Isekai Protagonist` class progression to have more Overpowered Abilities instead of training arc feats.
 - Changed `Glorious Aura` to give a +4 sacred bonus to all attributes instead.
 - Changed `Training Montage` to give a +10 untyped bonus to all attributes but now requires character level 10.
 - Changed `Extreme Speed` to give a +5 untyped speed bonus per character level.
-- Changed `Isekai Protagonist` class description.
 - Changed `Perfect Roll` and Auto metamagic abilities to be turned on by default.
+- Moved `Exceptional Summoning` to `Exceptional feats`.
 ### Removed
 - Removed `Winner` feature.
 - Removed `Training Arc` feature selection. Training Arc feats moved to character development.